Pythonのモデル周りを整理するために、ソースファイルを分割した
以下を参考に
https://qiita.com/shirakiya/items/0114d51e9c189658002e
